requirements.txt
# ============================================================
#  E-Commerce Churn Prediction — Dependencias de Producción
#  Generar con: pip freeze > requirements.txt
#  Instalar con: pip install -r requirements.txt
# ============================================================

# ─── Core científico ────────────────────────────────────────
numpy>=1.26.0,<2.0.0
pandas>=2.2.0,<3.0.0
pyarrow>=15.0.0          # Soporte Parquet

# ─── Machine Learning ───────────────────────────────────────
scikit-learn>=1.4.0,<2.0.0
xgboost>=2.0.0,<3.0.0
imbalanced-learn>=0.12.0  # SMOTE y variantes

# ─── Interpretabilidad ──────────────────────────────────────
shap>=0.45.0

# ─── Optimización de Hiperparámetros ────────────────────────
optuna>=3.5.0

# ─── Serialización de Modelos ───────────────────────────────
joblib>=1.3.0

# ─── Visualización ──────────────────────────────────────────
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# ─── Base de Datos / Supabase ───────────────────────────────
supabase>=2.3.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# ─── Validación de Datos ────────────────────────────────────
pandera>=0.19.0

# ─── Configuración ──────────────────────────────────────────
pyyaml>=6.0.1

# ─── Reportes ───────────────────────────────────────────────
jinja2>=3.1.3             # Templates para reportes HTML
nbconvert>=7.14.0         # Notebooks → HTML/PDF

# ─── Monitoreo de Drift ─────────────────────────────────────
evidently>=0.4.30

# ─── Utilidades ─────────────────────────────────────────────
tqdm>=4.66.0              # Barras de progreso
python-dateutil>=2.9.0